+++ lg_common/scripts/touchscreen.py
@@ -21,6 +21,7 @@ if __name__ == '__main__':
     url_base = rospy.get_param('~url_base', 'http://lg-head/ros_touchscreens/ts/')
     command_line_args = rospy.get_param('~command_line_args', '')
     extra_logging = rospy.get_param('~extra_logging', False)
+
     # TODO (wz) director_host and director_port should be global
 
     director_host = rospy.get_param('~director_host', '42-a')