+++ lg_keyboard/src/lg_keyboard/__init__.py
@@ -1,0 +1,5 @@
+from onboard import ROS_NODE_NAME
+from onboard import OnboardViewportException
+from onboard import OnboardLauncher
+from onboard import OnboardConfig
+from onboard_router import OnboardRouter