+++ liquidgalaxy/CHANGELOG.rst
@@ -2,6 +2,32 @@
 Changelog for package liquidgalaxy
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.4.12 (2016-10-07)
+-------------------
+
+1.4.11 (2016-10-06)
+-------------------
+
+1.4.10 (2016-10-06)
+-------------------
+
+1.4.9 (2016-10-04)
+------------------
+
+1.4.8 (2016-10-03)
+------------------
+* Update metapackage deps
+* Contributors: Matt Vollrath
+
+1.4.7 (2016-10-03)
+------------------
+* More changelogs
+* Generated changelog
+* Contributors: Wojciech Ziniewicz
+
+* Generated changelog
+* Contributors: Wojciech Ziniewicz
+
 1.4.6 (2016-09-28)
 ------------------